+++ src/client/components/evaluate/detail/EqualPopulation.tsx
@@ -1,41 +1,60 @@
 /** @jsx jsx */
-import { Box, Flex, jsx, ThemeUIStyleObject, Heading } from "theme-ui";
+import { Box, Flex, jsx, Styled, ThemeUIStyleObject, Heading } from "theme-ui";
 import { EvaluateMetricWithValue, DistrictProperties } from "../../../../shared/entities";
 import { getChoroplethStops } from "../../map/index";
 import { DistrictsGeoJSON } from "../../../types";
 
 const style: ThemeUIStyleObject = {
-  tableElement: {
+  table: {
+    mx: 0,
+    mb: 2,
+    width: "100%"
+  },
+  th: {
+    fontWeight: "bold",
+    color: "gray.7",
+    bg: "muted",
+    fontSize: 1,
+    textAlign: "left",
+    pt: 2,
+    px: 2,
+    height: "32px",
+    position: "sticky",
+    top: "0",
+    zIndex: 2,
+    userSelect: "none",
+    "&::after": {
+      height: "1px",
+      content: "''",
+      display: "block",
+      width: "100%",
+      bg: "gray.2",
+      bottom: "-1px",
+      position: "absolute",
+      left: 0,
+      right: 0
+    }
+  },
+  td: {
     fontWeight: "body",
     color: "gray.8",
     fontSize: 1,
     p: 2,
     textAlign: "left",
     verticalAlign: "bottom",
-    minWidth: "20px",
-    maxWidth: "200px",
-    mr: "10px"
+    position: "relative"
   },
-  fillBox: {
-    height: "10px",
-    width: "10px",
-    background: "orange",
-    opacity: "0.9",
-    outline: "1px solid black"
+  colFirst: {
+    pl: 0
   },
-  unfilledBox: {
-    height: "10px",
-    width: "10px",
-    background: "none",
-    outline: "1px solid black"
+  colLast: {
+    pr: 0
   },
-  compactnessItem: {
-    display: "inline-block",
-    verticalAlign: "center"
+  blankValue: {
+    color: "gray.2"
   },
-  tableHeader: {
-    float: "left",
-    width: "40%"
+  number: {
+    textAlign: "right"
   }
 };
 
@@ -67,61 +86,60 @@ const EqualPopulationMetricDetail = ({
   }
 
   return (
-    <Box sx={{ ml: "20px", overflowY: "scroll" }}>
-      <Heading as="h4" sx={{ variant: "text.h4", display: "block", width: "100%" }}>
+    <Box>
+      <Heading as="h2" sx={{ variant: "text.h5", mt: 4 }}>
         {metric.value || " "} of the {metric.total} districts are within{" "}
         {"popThreshold" in metric &&
           metric.popThreshold &&
           ` ${Math.floor(metric.popThreshold * 100)}%`}{" "}
         of {metric.avgPopulation && Math.floor(metric.avgPopulation).toLocaleString()}
       </Heading>
-      <Flex sx={{ flexDirection: "column", width: "60%" }}>
-        <table>
-          <thead>
-            <tr>
-              <th sx={style.tableElement}>Number</th>
-              <th sx={style.tableElement}>Deviation (%)</th>
-              <th sx={style.tableElement}>Deviation</th>
-            </tr>
-          </thead>
-          <tbody>
-            {geojson?.features.map(
-              (feature, id) =>
-                id > 0 && (
-                  <tr key={id}>
-                    <td sx={style.tableElement}>{id}</td>
-                    <td sx={style.tableElement}>
-                      <Box
-                        sx={{
-                          display: "inline-block",
-                          mr: "40px",
-                          width: "20px",
-                          height: "20px",
-                          opacity: "0.9",
-                          // @ts-ignore
-                          backgroundColor: feature.properties.percentDeviation
-                            ? `${computeRowFill(feature.properties)}`
-                            : "none"
-                        }}
-                      ></Box>
-                      <Box sx={style.compactnessItem}>
-                        {// @ts-ignore
-                        feature.properties.percentDeviation
-                          ? ` ${Math.floor(feature.properties.percentDeviation * 1000) / 10}%`
-                          : "-"}
-                      </Box>
-                    </td>
-                    <td sx={style.tableElement}>
-                      {feature.properties.populationDeviation
-                        ? Math.round(feature.properties.populationDeviation).toLocaleString()
-                        : "-"}
-                    </td>
-                  </tr>
-                )
-            )}
-          </tbody>
-        </table>
-      </Flex>
+      <Styled.table sx={style.table}>
+        <thead>
+          <Styled.tr>
+            <Styled.th sx={{ ...style.th, ...style.colFirst }}>Number</Styled.th>
+            <Styled.th sx={style.th}>Deviation (%)</Styled.th>
+            <Styled.th sx={{ ...style.th, ...style.number, ...style.colLast }}>Deviation</Styled.th>
+          </Styled.tr>
+        </thead>
+        <tbody>
+          {geojson?.features.map(
+            (feature, id) =>
+              id > 0 && (
+                <Styled.tr key={id}>
+                  <Styled.td sx={{ ...style.td, ...style.colFirst }}>{id}</Styled.td>
+
+                  <Styled.td sx={style.td}>
+                    {feature.properties.percentDeviation ? (
+                      <Flex sx={{ alignItems: "center" }}>
+                        <Styled.div
+                          sx={{
+                            mr: 2,
+                            width: "15px",
+                            height: "15px",
+                            borderRadius: "small",
+                            bg: computeRowFill(feature.properties)
+                          }}
+                        ></Styled.div>
+                        <Box>{Math.floor(feature.properties.percentDeviation * 1000) / 10}%</Box>
+                      </Flex>
+                    ) : (
+                      <Box sx={style.blankValue}>-</Box>
+                    )}
+                  </Styled.td>
+
+                  <Styled.td sx={{ ...style.td, ...style.number, ...style.colLast }}>
+                    {feature.properties.populationDeviation ? (
+                      Math.round(feature.properties.populationDeviation).toLocaleString()
+                    ) : (
+                      <Box sx={style.blankValue}>-</Box>
+                    )}
+                  </Styled.td>
+                </Styled.tr>
+              )
+          )}
+        </tbody>
+      </Styled.table>
     </Box>
   );
 };