+++ src/client/theme.ts
@@ -586,6 +586,32 @@ const theme: Theme & StyledSystemTheme = {
       alignItems: "center",
       flex: 1
     }
+  },
+  sidebar: {
+    white: {
+      bg: "muted",
+      boxShadow: "0 0 0 1px rgba(16,22,26,.1), 0 0 0 rgba(16,22,26,0), 0 1px 1px rgba(16,22,26,.2)",
+      display: "flex",
+      flexDirection: "column",
+      flexShrink: 0,
+      height: "100%",
+      minWidth: "400px",
+      position: "relative",
+      color: "gray.8",
+      zIndex: 200
+    },
+    gray: {
+      bg: "gray.0",
+      boxShadow: "0 0 0 1px rgba(16,22,26,.1), 0 0 0 rgba(16,22,26,0), 0 1px 1px rgba(16,22,26,.2)",
+      display: "flex",
+      flexDirection: "column",
+      flexShrink: 0,
+      height: "100%",
+      minWidth: "400px",
+      position: "relative",
+      color: "gray.8",
+      zIndex: 200
+    }
   }
 };